seed/python-sdk/unions/no-custom-config/requirements.txt
httpx>=0.21.2
pydantic>= 1.9.2
pydantic-core==2.18.2
typing_extensions>= 4.0.0
pytest-httpserver>= 1.1.3
pytest-asyncio
